add corrections
correction in edit item web
correction in header

+++ app/Repositories/ItemWebRepository.php
@@ -108,18 +108,21 @@ class ItemWebRepository
                 'section_web_id' => $data['section_web_id'],
                 'type_item_web_id' => $data['type_item_web_id'],
             ])){
-                // Assuming 'children' is a relationship, you would handle it like this
-                $itemWeb->children()->delete(); // Delete existing children
-                // Create new children without 'element_web' and 'item_web'
-                $cleanedChildren = [];
-                foreach ($data['children'] as $child) {
-                    // Remove 'element_web' and 'item_web' from each child data
-                    unset($child['id']);
-                    unset($child['element_web']);
-                    unset($child['item_web']);
-                    $cleanedChildren[] = $child;
+                if(isset($data['children'])){
+                    // Assuming 'children' is a relationship, you would handle it like this
+                    $itemWeb->children()->delete(); // Delete existing children
+                    // Create new children without 'element_web' and 'item_web'
+                    $cleanedChildren = [];
+                    foreach ($data['children'] as $child) {
+                        // Remove 'element_web' and 'item_web' from each child data
+                        unset($child['id']);
+                        unset($child['element_web']);
+                        unset($child['item_web']);
+                        $cleanedChildren[] = $child;
+                    }
+                    $itemWeb->children()->createMany($cleanedChildren);
                 }
-                $itemWeb->children()->createMany($cleanedChildren);
+              
                 if($itemWeb){
                     return $itemWeb;
                 }
